added location to the add tasks, and moved the location stuff to the main activity, also added to fragment
