+++ app/src/main/res/values/strings.xml
@@ -60,5 +60,6 @@
     <string name="image">image</string>
     <string name="detailimg">detailimg</string>
     <string name="fragmentimg">fragmentImg</string>
+    <string name="location">location</string>
     <!--    <string name="todo">TODO</string>-->
 </resources>
